[WB-25] PR feedback - rearrange middleware
