+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM node:8.4.0-alpine
+FROM node:8.6.0-alpine
 
 ENV HOME=/usr/windbreaker-hooks
